gcloud pub sub

+++ inferene_server/server_utils.py
@@ -4,6 +4,7 @@ import requests
 from threading import Thread
 from server_settings import BASE_DIR
 
+
 def webhook_response(training_webhook_url, data=None):
     def send(training_webhook_url, data=None):
         if training_webhook_url and "http" in training_webhook_url:
@@ -12,6 +13,7 @@ def webhook_response(training_webhook_url, data=None):
     Thread(target=send, args=(training_webhook_url, data)).start()
     return None
 
+
 def is_json_compatible(value):
     try:
         json.loads(value)
@@ -19,11 +21,15 @@ def is_json_compatible(value):
     except (TypeError, ValueError):
         return False
 
-def save_gcloud_storage_key():
-    env_var_name = "GCLOUD_STORAGE_KEY"
+
+def save_gcloud_keys(env_var_name: str, file_name: str):
+    file_name = os.path.join(BASE_DIR, file_name)
+    if os.path.exists(file_name):
+        print(f"The file {file_name} already exists.")
+        return
+    env_var_name = env_var_name
     env_var_value = os.getenv(env_var_name)
     if is_json_compatible(env_var_value):
-        file_name = os.path.join(BASE_DIR,"gcloud-storage-key.json")
         try:
             with open(file_name, "w") as json_file:
                 json.dump(json.loads(env_var_value), json_file, indent=4)
